Streamed the frozen second pass one file at a time
* ComputePass2TransferCompeteFull - the default protein-compact arm -
  takes the RescoredEntries seam instead of the whole-run pool and runs
  a file end to end: materialize, seed its pass-1 scalars, score with
  the frozen model, compete, stamp run q, write that file's
  .2nd-pass.fdr_scores.bin, drop it
* The sidecar is what carries the result forward, so the four
  experiment-scope columns - which the competition cannot know until
  every file has been read - are patched into those same files
  afterwards, per file, by the new
  FdrScoresSidecar.PatchExperimentValues. Its test asserts the contract
  that makes two phases substitutable for one: the patched file is
  byte-identical to a single-phase write. The map-back needs no entries
  at all, so it is a streamed rewrite rather than a second
  materialization
* Pass2SidecarWriter is now the one per-file write body. It existed
  twice, which is how the projection path acquired the --task
  ModelDiagnostics skip and the resident one did not; the resident
  block gains it here
* Pass1ScalarSeeder is RestorePass1Scalars per file, so a streamed
  caller can seed inside its own materialization. The whole-pool loop
  is skipped for the frozen modes, which would otherwise read every
  1st-pass sidecar twice and could only walk a resident pool
* The frozen competition modes got their own entry point rather than
  returning early out of ComputePass2Resident - nothing about them is
  resident any more
* Two deliberate changes on already-broken inputs: a duplicate
  --input-scores stem is last-wins rather than merged, since a streamed
  reader is handed one file's rows at a time; and a per-file key with no
  config.InputFiles entry now refuses, because such a file would take a
  fresh run q with nowhere to record its experiment q
* RescoredEntries.FileNames was set only by WithStreaming, which one of
  its three construction sites calls, so it was null on the resident
  rescore and the --task SecondPassFDR rehydrate. It now falls back to
  the buffer's keys the way Files() already did - free on those paths,
  which defer nothing. Nothing had read it before, and mode3 was the
  only leg that caught it
* Fixed six inspection warnings the score_index commits left: a
  ParquetScoreCache doc comment orphaned from ReadFdrStubScalars by an
  inserted method, and redundant casts in ReconciledParquetWriterTest

See TODO-20260826_osprey_stage7_stream_pool.md in pwiz-ai/todos

+++ pwiz_tools/Osprey/Osprey.Test/ReconciledParquetWriterTest.cs
@@ -143,14 +143,14 @@ namespace pwiz.Osprey.Test
             // Source: 3 rows for (100,2), 2 for (200,3) - five rows, ordinals 0..4.
             var sourceGroups = new List<(uint, byte, int)>
             {
-                (100u, (byte)2, 3),
-                (200u, (byte)3, 2),
+                (100u, 2, 3),
+                (200u, 3, 2),
             };
 
             // Same groups, same counts: every row pairs, ordinals in order. A rescored row
             // whose scan moved is still in its own group, so it pairs on position within it.
             var same = ParquetScoreCache.BuildScoreIndicesByPairing(
-                sourceGroups, new List<(uint, byte, int)> { (100u, (byte)2, 3), (200u, (byte)3, 2) },
+                sourceGroups, new List<(uint, byte, int)> { (100u, 2, 3), (200u, 3, 2) },
                 out int srcRows);
             Assert.AreEqual(5, srcRows);
             CollectionAssert.AreEqual(new uint[] { 0, 1, 2, 3, 4 }, same);
@@ -160,7 +160,7 @@ namespace pwiz.Osprey.Test
             // source row count, and the groups after it keep their own ordinals.
             var withGapGroup = ParquetScoreCache.BuildScoreIndicesByPairing(
                 sourceGroups,
-                new List<(uint, byte, int)> { (100u, (byte)2, 3), (150u, (byte)2, 1), (200u, (byte)3, 2) },
+                new List<(uint, byte, int)> { (100u, 2, 3), (150u, 2, 1), (200u, 3, 2) },
                 out _);
             CollectionAssert.AreEqual(new uint[] { 0, 1, 2, 5, 3, 4 }, withGapGroup);
 
@@ -168,7 +168,7 @@ namespace pwiz.Osprey.Test
             // never be handed an ordinal that belongs to a different Stage 4 row.
             var withExtraRow = ParquetScoreCache.BuildScoreIndicesByPairing(
                 sourceGroups,
-                new List<(uint, byte, int)> { (100u, (byte)2, 4), (200u, (byte)3, 2) },
+                new List<(uint, byte, int)> { (100u, 2, 4), (200u, 3, 2) },
                 out _);
             CollectionAssert.AreEqual(new uint[] { 0, 1, 2, 5, 3, 4 }, withExtraRow);
         }
